# backend/doc/openapi-docs.json
{
  "components": {
...
    "securitySchemes": {
      "bearer-key": {
        "bearerFormat": "JWT",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "title": "Rebus backend",
    "version": "4.0.0"
  },
  "openapi": "3.0.1",
  "paths": {
...
    "/pictures/{challengeId}/{userId}": {
      "post": {
        "operationId": "uploadPicture",
        "parameters": [
          {
            "in": "path",
            "name": "challengeId",
            "required": true,
            "schema": {
              "format": "int32",
              "type": "integer"
            }
          },
          {
            "in": "path",
            "name": "userId",
            "required": true,
            "schema": {
              "format": "int32",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "properties": {
                  "fileContent": {
                    "format": "binary",
                    "type": "string"
                  }
                },
                "required": ["fileContent"],
                "type": "object"
              }
            }
          }
        },
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "example": {
                    "data": "https://example.com/pictures/123/456",
                    "message": "",
                    "status": "SUCCESS"
                  },
                  "type": "string"
                }
              }
            },
            "description": "OK, Image URL in the body"
          },
          "400": {
            "content": {
              "*/*": {
                "schema": {
                  "example": {
                    "data": "",
                    "message": "Could not store image",
                    "status": "ERROR"
                  },
                  "type": "string"
                }
              }
            },
            "description": "Bad request, something went wrong with storing the image"
          },
          "403": {
            "content": {
              "*/*": {
                "schema": {
                  "example": {
                    "data": "",
                    "message": "Not allowed to upload images for other teams",
                    "status": "ERROR"
                  },
                  "type": "string"
                }
              }
            },
            "description": "Forbidden, not allowed to upload images for other teams"
          }
        },
        "summary": "Upload an image to the server",
        "tags": ["User endpoints"]
      }
    },
...
  },
  "security": [
    {
      "bearer-key": []
    }
  ],
  "servers": [
    {
      "description": "Generated server url",
      "url": "http://localhost:8080"
    }
  ]
}